pyproject.toml
[project]
name = "mlops-examples"
description = "Examples and tutorials for using mlops-services with DVC and MLflow"
requires-python = ">=3.11,<3.14"
version = "0.1.0"
dependencies = [
  "mlflow>=2.10,<3",
  "DVC[s3]>=3.0,<4",
  "boto3>=1.28",
  "scikit-learn>=1.3",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "matplotlib>=3.7",
]
